1、将hsweb-web-core模块session包中的siample包名修正为simple 2、将hsweb-web-core模块session.simple.SimpleHttpSessionManager.getSessionIdByUserId方法中的return user.getId修正为return session.getId()
